add some error handling

+++ src/products/products.controller.ts
@@ -1,10 +1,12 @@
 import {
+  BadRequestException,
   Body,
   Controller,
   Delete,
   Get,
   HttpCode,
   HttpStatus,
+  NotFoundException,
   Param,
   Post,
   Put,
@@ -23,13 +25,22 @@ export class ProductsController {
   constructor(private productsService: ProductsService) {}
 
   @Get()
-  getProducts(@Query() query: GetProductsDto) {
-    return this.productsService.findAll(query);
+  async getProducts(@Query() query: GetProductsDto) {
+    const products = await this.productsService.findAll(query);
+    if (!products) throw new NotFoundException('There are no any products.');
+
+    return products;
   }
 
   @Get('/:id')
-  getOneProduct(@Param('id') id: string) {
-    return this.productsService.findOne(+id);
+  async getOneProduct(@Param('id') id: string) {
+    if (!+id) throw new BadRequestException('Invalid id property');
+
+    const product = await this.productsService.findOne(+id);
+    if (!product)
+      throw new NotFoundException('There is no a product with such an id');
+
+    return product;
   }
 
   @Post()
@@ -39,12 +50,16 @@ export class ProductsController {
 
   @Put('/:id')
   updateProduct(@Param('id') id: string, @Body() body: UpdateProductDto) {
+    if (!+id) throw new BadRequestException('Invalid id property');
+
     return this.productsService.update(+id, body);
   }
 
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
   deleteProduct(@Param('id') id: string) {
+    if (!+id) throw new BadRequestException('Invalid id property');
+
     return this.productsService.delete(+id);
   }
 }
